+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/AnyExtensions.kt
@@ -24,7 +24,7 @@ fun Any?.smartToString(deepOperation: Boolean = true) = when {
 }
 
 
-/**将当期对象转化为字符串。如果为null，则转化为空字符串。*/
+/**将当前对象转化为字符串。如果为null，则转化为空字符串。*/
 fun Any?.toStringOrEmpty(): String = this?.toString() ?: ""