+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/CloneableExtensionsTest.kt
@@ -7,8 +7,10 @@ class CloneableExtensionsTest {
 
 	class DeepObj : Cloneable {
 		var a = 1
-		var b = arrayOf(1)
+		var b = intArrayOf(1)
 		var c = listOf(1)
+		var d = listOf(Obj())
+		var e = mutableListOf(Obj())
 		var obj: Obj = Obj()
 	}
 
@@ -20,11 +22,12 @@ class CloneableExtensionsTest {
 		val deepObj1 = DeepObj()
 		val deepObj2 = deepObj1.deepClone()
 
-		assertNotEquals(obj1, obj2)
-		assertNotEquals(deepObj1, deepObj2)
-		assertNotEquals(deepObj1.b, deepObj2.b)
-		assertNotEquals(deepObj1.c, deepObj2.c)
-		assertNotEquals(deepObj1.obj, deepObj2.obj)
+		assertNotSame(obj1, obj2)
+		assertNotSame(deepObj1, deepObj2)
+		assertNotSame(deepObj1.c, deepObj2.c)
+		assertNotSame(deepObj1.d, deepObj2.d)
+		assertNotSame(deepObj1.e, deepObj2.e)
+		assertNotSame(deepObj1.obj, deepObj2.obj)
 	}
 }